Update Lorem ipsum - plea for Dolerem ipsum.md
Ah, behold the elusive "Dolorem ipsum" from Cicero, a gem of linguistic finesse, seldom found amidst the cacophony of "Lorem ipsum" clones that populate the vast expanse of the internet. Alas, like a hidden treasure shrouded in obscurity, few dare to venture into the realm of ancient wisdom and instead settle for the tiresome and mundane familiarity of "Lorem ipsum." Verily, it seems the majestic words of Cicero are reserved for the enlightened few, while the masses content themselves with the commonplace. Oh, the irony of our modern age, where brilliance lies unnoticed amidst the commonplace drudgery of filler text!